move space check into generic
