+++ octopus-server/src/main/java/octopus/teamcity/server/generic/BuildInformationStep.java
@@ -21,8 +21,7 @@ public class BuildInformationStep extends OctopusBuildStep {
   }
 
   @Override
-  protected List<InvalidProperty> validateBuildSpecificProperties(
-      final Map<String, String> properties) {
+  public List<InvalidProperty> validateProperties(final Map<String, String> properties) {
     final List<InvalidProperty> failedProperties = Lists.newArrayList();
 
     final String packageId = properties.getOrDefault(KEYS.getPackageIdPropertyName(), "");
