+++ octopus-server/src/main/java/octopus/teamcity/server/generic/PushPackageStep.java
@@ -20,8 +20,7 @@ public class PushPackageStep extends OctopusBuildStep {
   }
 
   @Override
-  protected List<InvalidProperty> validateBuildSpecificProperties(
-      final Map<String, String> properties) {
+  public List<InvalidProperty> validateProperties(final Map<String, String> properties) {
     final List<InvalidProperty> failedProperties = Lists.newArrayList();
 
     final String packagePaths = properties.getOrDefault(KEYS.getPackagePathsPropertyName(), "");
