+++ octopus-server/src/test/java/octopus/teamcity/server/generic/RunbookRunStepTest.java
@@ -8,7 +8,6 @@ import java.util.Map;
 import java.util.stream.Stream;
 
 import jetbrains.buildServer.serverSide.InvalidProperty;
-import octopus.teamcity.common.commonstep.CommonStepPropertyNames;
 import octopus.teamcity.common.runbookrun.RunbookRunPropertyNames;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -71,7 +70,6 @@ class RunbookRunStepTest {
   private Map<String, String> buildPropertiesMap() {
     final Map<String, String> validMap = new HashMap<>();
     // Mandatory/validated
-    validMap.put(CommonStepPropertyNames.SPACE_NAME, "TheSpace");
     validMap.put(RunbookRunPropertyNames.RUNBOOK_NAME, "RunbookName");
     validMap.put(RunbookRunPropertyNames.PROJECT_NAME, "ProjectName");
     validMap.put(RunbookRunPropertyNames.ENVIRONMENT_NAMES, "Env1\nEnv2");
@@ -111,10 +109,6 @@ class RunbookRunStepTest {
         Arguments.of(
             RunbookRunPropertyNames.ENVIRONMENT_NAMES,
             "env1\n \nenv3",
-            "An environment name cannot be whitespace."),
-        Arguments.of(
-            CommonStepPropertyNames.SPACE_NAME,
-            null,
-            "Space name must be specified, and cannot be whitespace."));
+            "An environment name cannot be whitespace."));
   }
 }
